pyproject.toml
[tool.poetry]
name = "aqua-parser"
version = "0.1.0"
description = ""
authors = ["astra <zorkin.roman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
tabula-py = "^2.3.0"
pdfplumber = "^0.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.941"
# fix bandit 1.7.3-4 errors
bandit = "1.7.2"
wemake-python-styleguide = "^0.16.1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
